Set project to allow automatic loading of default files
The oslo_config project allows some automatic discovery of config files,
but that needs the project to be set.
So, we try first to parse the component from the arguments with
argparse, and failing that, guessing it from the host-name.

If that fails, we fall back to the old behaviour, so nothing is lost.