Support caching the token
Since `openstack-agent-liveness` is a short-running command, it cannot
hold the token and session for talking to OpenStack for a long time.
With being run very often in a lot of pods, this leads to a lot of
requests to Keystone to request a new token. Since these tokens are new,
they are also not known to the appropriate service
`openstack-agent-liveness` wants to talk to. Therefore, said service
needs to make another request against Keystone to validate the token
before allowing our request.

To overcome this, we implement token caching. If this functionality is
enabled by passing the path to a file used for the cache with the
`--token_cache_file` CLI option, communication with Keystone only
happens if the token expired or is about to expire. Starting with an
empty cache or a non-existent file is handled gracefully by falling back
to password authentication for the first request.